share paginateMyPupilsHandler between updatePresentationState and GetMyPupils so that the CurrentPageOfPupils is not stored in the form
