--- DfE.GIAP.All/src/DfE.GIAP.Web/Controllers/MyPupilList/Services/Presentation/PupilDtoPresentationHandlers/Options/IPupilPresentationOptionsProvider.cs
@@ -1,7 +1,0 @@
-﻿namespace DfE.GIAP.Web.Controllers.MyPupilList.Services.Presentation.PupilDtoPresentationHandlers.Options;
-
-public interface IPupilsPresentationOptionsProvider
-{
-    PupilsPresentationOptions GetOptions();
-    void SetOptions(PupilsPresentationOptions options);
-}
