--- DfE.GIAP.All/src/DfE.GIAP.Web/Controllers/MyPupilList/Services/Presentation/PupilDtoPresentationHandlers/Options/PupilPresentationOptions.cs
@@ -1,6 +1,0 @@
-﻿namespace DfE.GIAP.Web.Controllers.MyPupilList.Services.Presentation.PupilDtoPresentationHandlers.Options;
-
-public record PupilsPresentationOptions(
-    int Page,
-    string SortBy,
-    SortDirection SortDirection);
