--- DfE.GIAP.All/src/DfE.GIAP.Web/Controllers/MyPupilList/Services/Presentation/PupilDtoPresentationHandlers/Options/SortDirection.cs
@@ -1,7 +1,0 @@
-﻿namespace DfE.GIAP.Web.Controllers.MyPupilList.Services.Presentation.PupilDtoPresentationHandlers.Options;
-
-public enum SortDirection
-{
-    Ascending,
-    Descending
-}
