--- DfE.GIAP.All/src/DfE.GIAP.Web/Controllers/MyPupilList/Services/Presentation/PupilDtoPresentationHandlers/PaginateHandler/PaginatePupilDtosPresentationHandler.cs
@@ -1,26 +1,0 @@
-﻿using DfE.GIAP.Core.MyPupils.Application.UseCases.GetMyPupils.Response;
-using DfE.GIAP.Web.Controllers.MyPupilList.Services.Presentation.PupilDtoPresentationHandlers.Options;
-
-namespace DfE.GIAP.Web.Controllers.MyPupilList.Services.Presentation.PupilDtoPresentationHandlers.PaginateHandler;
-
-public sealed class PaginatePupilDtosPresentationHandler : IPupilDtoPresentationHandler
-{
-    public IEnumerable<PupilDto> Handle(IEnumerable<PupilDto> pupils, PupilsPresentationOptions options)
-    {
-        const int DefaultPageSize = 20;
-        PageNumber page = PageNumber.Page(options.Page);
-        int skip = DefaultPageSize * (page.Value - 1);
-
-        if (skip >= pupils.Count())
-        {
-            return [];
-        }
-
-        List<PupilDto> pagedResults = pupils
-            .Skip(skip)
-            .Take(DefaultPageSize)
-            .ToList();
-
-        return pagedResults;
-    }
-}
