--- DfE.GIAP.All/src/DfE.GIAP.Web/Controllers/MyPupilList/Services/Presentation/PupilSelectionState/Provider/IPupilSelectionStateProvider.cs
@@ -1,7 +1,0 @@
-﻿namespace DfE.GIAP.Web.Controllers.MyPupilList.Services.Presentation.PupilSelectionState.Provider;
-
-public interface IPupilSelectionStateProvider
-{
-    PupilsSelectionState GetState();
-    void UpdateState(PupilsSelectionState state);
-}
