--- DfE.GIAP.All/src/DfE.GIAP.Web/Controllers/MyPupilList/Services/Presentation/PupilSelectionState/SelectAllPupilsState.cs
@@ -1,8 +1,0 @@
-﻿namespace DfE.GIAP.Web.Controllers.MyPupilList.Services.Presentation.PupilSelectionState;
-
-public enum SelectAllPupilsState
-{
-    SelectAll,
-    DeselectAll,
-    NotSpecified
-}
