+++ DfE.GIAP.All/src/DfE.GIAP.Web/Features/MyPupils/Handlers/GetMyPupils/Extensions/SortDirectionExtensions.cs
@@ -1,0 +1,16 @@
+﻿using DfE.GIAP.Web.Features.MyPupils.PresentationState;
+
+namespace DfE.GIAP.Web.Features.MyPupils.Handlers.GetMyPupils.Extensions;
+
+public static class SortDirectionExtensions
+{
+    public static string ToFormSortDirection(this SortDirection sortDirection)
+    {
+        return sortDirection switch
+        {
+            SortDirection.Ascending => "asc",
+            SortDirection.Descending => "desc",
+            _ => string.Empty
+        };
+    }
+}
