+++ DfE.GIAP.All/src/DfE.GIAP.Web/Features/MyPupils/Handlers/GetMyPupils/IGetMyPupilsHandler.cs
@@ -1,0 +1,8 @@
+﻿using DfE.GIAP.Web.Features.MyPupils.ViewModels;
+
+namespace DfE.GIAP.Web.Features.MyPupils.Handlers.GetMyPupils;
+
+public interface IGetMyPupilsHandler
+{
+    Task<IEnumerable<PupilViewModel>> HandleAsync(GetMyPupilsRequest request);
+}
