+++ DfE.GIAP.All/src/DfE.GIAP.Web/Features/MyPupils/Handlers/GetPaginatedMyPupils/GetPaginatedMyPupilsRequest.cs
@@ -1,0 +1,5 @@
+﻿using DfE.GIAP.Web.Features.MyPupils.PresentationState;
+
+namespace DfE.GIAP.Web.Features.MyPupils.Handlers.GetPaginatedMyPupils;
+
+public record GetPaginatedMyPupilsRequest(string UserId, MyPupilsPresentationState PresentationState);
