+++ DfE.GIAP.All/src/DfE.GIAP.Web/Features/MyPupils/Handlers/GetPaginatedMyPupils/IGetPaginatedMyPupilsHandler.cs
@@ -1,0 +1,8 @@
+﻿using DfE.GIAP.Core.MyPupils.Application.UseCases.GetMyPupils.Response;
+
+namespace DfE.GIAP.Web.Features.MyPupils.Handlers.GetPaginatedMyPupils;
+
+public interface IGetPaginatedMyPupilsHandler
+{
+    Task<IEnumerable<PupilDto>> GetPaginatedPupilsAsync(GetPaginatedMyPupilsRequest request);
+}
