+++ DfE.GIAP.All/src/DfE.GIAP.Web/Features/MyPupils/Handlers/GetPaginatedMyPupils/PresentationHandlers/Paginate/PageNumber.cs
@@ -1,0 +1,13 @@
+﻿namespace DfE.GIAP.Web.Features.MyPupils.Handlers.GetPaginatedMyPupils.PresentationHandlers.Paginate;
+
+public readonly struct PageNumber
+{
+    public PageNumber(int page)
+    {
+        ArgumentOutOfRangeException.ThrowIfLessThan(page, 1);
+        Value = page;
+    }
+
+    public int Value { get; }
+    public static PageNumber Page(int page) => new(page);
+}
