+++ DfE.GIAP.All/src/DfE.GIAP.Web/Features/MyPupils/Handlers/UpdateMyPupilsState/IUpdateMyPupilsStateHandler.cs
@@ -1,0 +1,7 @@
+﻿namespace DfE.GIAP.Web.Features.MyPupils.Handlers.UpdateMyPupilsState;
+
+public interface IUpdateMyPupilsStateHandler
+{
+    Task HandleAsync(UpdateMyPupilsStateRequest request);
+}
+
