+++ DfE.GIAP.All/src/DfE.GIAP.Web/Features/MyPupils/Handlers/UpdateMyPupilsState/UpdateMyPupilsStateRequest.cs
@@ -1,0 +1,9 @@
+﻿using DfE.GIAP.Web.Features.MyPupils;
+using DfE.GIAP.Web.Features.MyPupils.PresentationState;
+
+namespace DfE.GIAP.Web.Features.MyPupils.Handlers.UpdateMyPupilsState;
+
+public record UpdateMyPupilsStateRequest(
+    string UserId,
+    MyPupilsPresentationState CurrentPresentationState,
+    MyPupilsFormStateRequestDto UpdateStateInput);
