+++ DfE.GIAP.All/src/DfE.GIAP.Web/Features/MyPupils/PresentationState/MyPupilsPresentationState.cs
@@ -1,0 +1,6 @@
+﻿namespace DfE.GIAP.Web.Features.MyPupils.PresentationState;
+
+public record MyPupilsPresentationState(
+    int Page,
+    string SortBy,
+    SortDirection SortDirection);
