+++ DfE.GIAP.All/src/DfE.GIAP.Web/Features/MyPupils/PresentationState/Provider/IMyPupilsPresentationStateProvider.cs
@@ -1,0 +1,9 @@
+﻿using DfE.GIAP.Web.Features.MyPupils.PresentationState;
+
+namespace DfE.GIAP.Web.Features.MyPupils.PresentationState.Provider;
+
+public interface IMyPupilsPresentationStateProvider
+{
+    MyPupilsPresentationState GetState();
+    void SetState(MyPupilsPresentationState state);
+}
