+++ DfE.GIAP.All/src/DfE.GIAP.Web/Features/MyPupils/PresentationState/SortDirection.cs
@@ -1,0 +1,7 @@
+﻿namespace DfE.GIAP.Web.Features.MyPupils.PresentationState;
+
+public enum SortDirection
+{
+    Ascending,
+    Descending
+}
