+++ DfE.GIAP.All/src/DfE.GIAP.Web/Features/MyPupils/SelectionState/Provider/DataTransferObjects/MyPupilsPupilSelectionStateDto.cs
@@ -1,0 +1,7 @@
+﻿namespace DfE.GIAP.Web.Features.MyPupils.SelectionState.Provider.DataTransferObjects;
+
+public sealed class MyPupilsPupilSelectionStateDto
+{
+    public Dictionary<string, bool> PupilUpnToSelectedMap { get; set; } = [];
+    public PupilSelectionModeDto State { get; set; } = PupilSelectionModeDto.NotSpecified;
+}
