+++ DfE.GIAP.All/src/DfE.GIAP.Web/Features/MyPupils/SelectionState/Provider/DataTransferObjects/PupilSelectionModeDto.cs
@@ -1,0 +1,8 @@
+﻿namespace DfE.GIAP.Web.Features.MyPupils.SelectionState.Provider.DataTransferObjects;
+
+public enum PupilSelectionModeDto
+{
+    SelectAll,
+    DeselectAll,
+    NotSpecified
+}
