+++ DfE.GIAP.All/src/DfE.GIAP.Web/Features/MyPupils/SelectionState/Provider/IMyPupilsPupilSelectionStateProvider.cs
@@ -1,0 +1,9 @@
+﻿using DfE.GIAP.Web.Features.MyPupils.SelectionState;
+
+namespace DfE.GIAP.Web.Features.MyPupils.SelectionState.Provider;
+
+public interface IMyPupilsPupilSelectionStateProvider
+{
+    MyPupilsPupilSelectionState GetState();
+    void SetState(MyPupilsPupilSelectionState state);
+}
