+++ DfE.GIAP.All/src/DfE.GIAP.Web/Features/MyPupils/ViewModels/MyPupilsErrorViewModel.cs
@@ -1,0 +1,3 @@
+﻿namespace DfE.GIAP.Web.Features.MyPupils.ViewModels;
+
+public record MyPupilsErrorViewModel(string Message); // TODO add behaviour
