+++ DfE.GIAP.All/tests/DfE.GIAP.Web.Tests/Controllers/MyPupilList/PupilsSelectionStateTests.cs
@@ -1,6 +1,5 @@
 ﻿using DfE.GIAP.SharedTests.TestDoubles;
-using DfE.GIAP.Web.Controllers.MyPupilList.Services.Presentation.PupilSelectionState;
-using DfE.GIAP.Web.Controllers.MyPupilList.Services.Presentation.PupilSelectionState.Provider.DataTransferObjects;
+using DfE.GIAP.Web.Features.MyPupils.SelectionState;
 using DfE.GIAP.Web.Tests.Controllers.MyPupilList.TestDoubles;
 using Moq;
 using Xunit;
@@ -12,25 +11,25 @@ public sealed class PupilSelectionStateTests
     public void Default_State_Is_Empty()
     {
         // Arrange Act
-        PupilsSelectionState state = PupilsSelectionStateTestDoubles.CreateEmpty();
+        MyPupilsPupilSelectionState state = PupilsSelectionStateTestDoubles.CreateEmpty();
 
         // Assert
         Assert.False(state.IsAllPupilsSelected);
-        Assert.Empty(state.GetSelectedPupils());
+        Assert.Empty(state.GetPupilsWithSelectionState());
     }
 
     [Fact]
     public void SelectAllPupils_Updates_State_WithEmptyPupils()
     {
         // Arrange
-        PupilsSelectionState state = PupilsSelectionStateTestDoubles.CreateEmpty();
+        MyPupilsPupilSelectionState state = PupilsSelectionStateTestDoubles.CreateEmpty();
 
         // Act
         state.SelectAllPupils();
 
         // Assert
         Assert.True(state.IsAllPupilsSelected);
-        Assert.Empty(state.GetSelectedPupils());
+        Assert.Empty(state.GetPupilsWithSelectionState());
     }
 
     [Fact]
@@ -38,14 +37,13 @@ public sealed class PupilSelectionStateTests
     {
         // Arrange
         List<string> upns = UniquePupilNumberTestDoubles.GenerateAsValues(count: 3);
-        PupilsSelectionState state = PupilsSelectionStateTestDoubles.CreateWithPupilUniquePupilNumbers(upns);
+        MyPupilsPupilSelectionState state = PupilsSelectionStateTestDoubles.CreateWithSelectionState(upns);
 
         // Act
         state.SelectAllPupils();
 
         // Assert
         Assert.True(state.IsAllPupilsSelected);
-        Assert.Equivalent(upns, state.GetSelectedPupils());
         Assert.All(upns, upn => Assert.True(state.IsPupilSelected(upn)));
     }
 
@@ -54,17 +52,18 @@ public sealed class PupilSelectionStateTests
     {
         // Arrange
         List<string> upns = UniquePupilNumberTestDoubles.GenerateAsValues(count: 3);
-        PupilsSelectionState state = PupilsSelectionStateTestDoubles.CreateWithPupilUniquePupilNumbers(upns);
+        MyPupilsPupilSelectionState state = PupilsSelectionStateTestDoubles.CreateWithSelectionState(upns);
 
         // Act
-        state.UpdatePupilSelectionState([upns[0]], true);
-        state.UpdatePupilSelectionState([upns[1]], false);
+        state.UpsertPupilSelectionState([upns[0]], true);
+        state.UpsertPupilSelectionState([upns[1]], false);
 
         // Assert
         Assert.True(state.IsPupilSelected(upns[0]));
         Assert.False(state.IsPupilSelected(upns[1]));
-        Assert.Contains(upns[0], state.GetSelectedPupils());
-        Assert.DoesNotContain(upns[1], state.GetSelectedPupils());
+        IReadOnlyDictionary<string, bool> selectionState = state.GetPupilsWithSelectionState();
+        Assert.Contains(upns[0], selectionState);
+        Assert.Contains(upns[1], selectionState);
     }
 
     [Fact]
@@ -72,7 +71,7 @@ public sealed class PupilSelectionStateTests
     {
         // Arrange
         List<string> upns = UniquePupilNumberTestDoubles.GenerateAsValues(count: 3);
-        PupilsSelectionState state = PupilsSelectionStateTestDoubles.CreateWithPupilUniquePupilNumbers(upns);
+        MyPupilsPupilSelectionState state = PupilsSelectionStateTestDoubles.CreateWithSelectionState(upns);
 
         // Act
         state.SelectAllPupils();
@@ -80,7 +79,7 @@ public sealed class PupilSelectionStateTests
 
         // Assert
         Assert.False(state.IsAllPupilsSelected);
-        Assert.Empty(state.GetSelectedPupils());
+        Assert.Equal(3, state.GetPupilsWithSelectionState().Count);
         Assert.All(upns, upn => Assert.False(state.IsPupilSelected(upn)));
     }
 
@@ -89,16 +88,16 @@ public sealed class PupilSelectionStateTests
     {
         // Arrange
         List<string> upns = UniquePupilNumberTestDoubles.GenerateAsValues(count: 2);
-        PupilsSelectionState state = PupilsSelectionStateTestDoubles.CreateWithPupilUniquePupilNumbers(upns);
+        MyPupilsPupilSelectionState state = PupilsSelectionStateTestDoubles.CreateWithSelectionState(upns);
         string selectedUpn = upns[0];
-        state.UpdatePupilSelectionState([selectedUpn], true);
+        state.UpsertPupilSelectionState([selectedUpn], true);
 
         // Act
         state.ResetState();
 
         // Assert
         Assert.False(state.IsAllPupilsSelected);
-        Assert.Empty(state.GetSelectedPupils());
+        Assert.Empty(state.GetPupilsWithSelectionState());
         Assert.All(upns, upn => Assert.False(state.IsPupilSelected(upn)));
     }
 
@@ -106,22 +105,22 @@ public sealed class PupilSelectionStateTests
     public void AddPupils_WithNullOrEmpty_Throws()
     {
         // Arrange Act
-        PupilsSelectionState state = PupilsSelectionStateTestDoubles.CreateEmpty();
+        MyPupilsPupilSelectionState state = PupilsSelectionStateTestDoubles.CreateEmpty();
 
         // Assert
-        Assert.Throws<ArgumentNullException>(() => state.AddPupils(null));
-        Assert.Throws<ArgumentException>(() => state.AddPupils([]));
+        Assert.Throws<ArgumentNullException>(() => state.UpsertPupilSelectionState(null, It.IsAny<bool>()));
+        Assert.Throws<ArgumentException>(() => state.UpsertPupilSelectionState([], It.IsAny<bool>()));
     }
 
     [Fact]
     public void AddPupils_WithInvalidUpn_Throws()
     {
         // Arrange
-        PupilsSelectionState state = PupilsSelectionStateTestDoubles.CreateEmpty();
+        MyPupilsPupilSelectionState state = PupilsSelectionStateTestDoubles.CreateEmpty();
         List<string> invalidUpns = ["INVALID_UPN"];
 
         // Act & Assert
-        ArgumentException ex = Assert.Throws<ArgumentException>(() => state.AddPupils(invalidUpns));
+        ArgumentException ex = Assert.Throws<ArgumentException>(() => state.UpsertPupilSelectionState(invalidUpns, It.IsAny<bool>()));
         Assert.Equal("Invalid UPN requested", ex.Message);
     }
 
@@ -130,15 +129,14 @@ public sealed class PupilSelectionStateTests
     public void AddPupils_WithDuplicateUpns_DoesNotThrowOrDuplicate()
     {
         // Arrange
-        PupilsSelectionState state = PupilsSelectionStateTestDoubles.CreateEmpty();
+        MyPupilsPupilSelectionState state = PupilsSelectionStateTestDoubles.CreateEmpty();
         string upn = UniquePupilNumberTestDoubles.Generate().Value;
-        state.AddPupils([upn, upn]);
 
         // Act
-        state.UpdatePupilSelectionState([upn], true);
+        state.UpsertPupilSelectionState([upn, upn], true);
 
         // Assert
-        Assert.Single(state.GetSelectedPupils());
+        Assert.Single(state.GetPupilsWithSelectionState());
         Assert.True(state.IsPupilSelected(upn));
     }
 
@@ -147,89 +145,53 @@ public sealed class PupilSelectionStateTests
     public void UpdateSelectionState_OnUnknownUpn_AddsAndMarksSelected()
     {
         // Arrange
-        PupilsSelectionState state = PupilsSelectionStateTestDoubles.CreateEmpty();
+        MyPupilsPupilSelectionState state = PupilsSelectionStateTestDoubles.CreateEmpty();
         string upn = UniquePupilNumberTestDoubles.Generate().Value;
 
         // Act
-        state.UpdatePupilSelectionState([upn], true);
+        state.UpsertPupilSelectionState([upn], true);
 
         // Assert
         Assert.True(state.IsPupilSelected(upn));
-        Assert.Contains(upn, state.GetSelectedPupils());
+        Assert.True(state.GetPupilsWithSelectionState()[upn]);
     }
 
     [Fact]
     public void UpdateSelectionState_OnUnknownUpn_AddsAndMarksDeselected()
     {
         // Arrange
-        PupilsSelectionState state = PupilsSelectionStateTestDoubles.CreateEmpty();
+        MyPupilsPupilSelectionState state = PupilsSelectionStateTestDoubles.CreateEmpty();
         string upn = UniquePupilNumberTestDoubles.Generate().Value;
 
         // Act
-        state.UpdatePupilSelectionState([upn], false);
+        state.UpsertPupilSelectionState([upn], false);
 
 
         // Assert
         Assert.False(state.IsPupilSelected(upn));
-        Assert.DoesNotContain(upn, state.GetSelectedPupils());
+        Assert.False(state.GetPupilsWithSelectionState()[upn]);
     }
 
     [Fact]
     public void UpdateSelectionState_WithInvalidUpn_Throws()
     {
         // Arrange
-        PupilsSelectionState state = PupilsSelectionStateTestDoubles.CreateEmpty();
+        MyPupilsPupilSelectionState state = PupilsSelectionStateTestDoubles.CreateEmpty();
         List<string> invalidUpns = ["INVALID-1", "INVALID-2"];
 
         // Act & Assert
-        Assert.Throws<ArgumentException>(() => state.UpdatePupilSelectionState(invalidUpns, It.IsAny<bool>()));
+        Assert.Throws<ArgumentException>(() => state.UpsertPupilSelectionState(invalidUpns, It.IsAny<bool>()));
     }
 
     [Fact]
     public void UpdateSelectionState_WithMixedValidAndInvalidUpns_Throws()
     {
         // Arrange
-        PupilsSelectionState state = PupilsSelectionStateTestDoubles.CreateEmpty();
+        MyPupilsPupilSelectionState state = PupilsSelectionStateTestDoubles.CreateEmpty();
         string validUpn = UniquePupilNumberTestDoubles.Generate().Value;
         List<string> mixedUpns = [validUpn, "INVALID_UPN"];
 
         // Act & Assert
-        Assert.Throws<ArgumentException>(() => state.UpdatePupilSelectionState(mixedUpns, It.IsAny<bool>()));
-    }
-
-    [Fact]
-    public void ToDto_And_FromDto_Preserves_State()
-    {
-        // Arrange
-        List<string> upns = UniquePupilNumberTestDoubles.GenerateAsValues(count: 2);
-        PupilsSelectionState original = PupilsSelectionStateTestDoubles.CreateEmpty();
-        original.AddPupils(upns);
-        original.UpdatePupilSelectionState([upns[0]], true);
-        original.SelectAllPupils();
-
-        // Act
-        PupilSelectionStateDto dto = original.ToDto();
-        PupilsSelectionState restored = PupilsSelectionState.FromDto(dto);
-
-        // Assert
-        Assert.True(restored.IsAllPupilsSelected);
-        Assert.Equivalent(original.GetSelectedPupils(), restored.GetSelectedPupils());
-        Assert.All(upns, upn => Assert.True(restored.IsPupilSelected(upn)));
-    }
-
-    [Fact]
-    public void FromDto_WithEmptyMapAndUnspecifiedState_CreatesEmptyState()
-    {
-        // Arrange
-        PupilSelectionStateDto dto = new()
-        {
-            PupilUpnToSelectedMap = [],
-            State = SelectAllPupilsState.NotSpecified
-        };
-
-        PupilsSelectionState state = PupilsSelectionState.FromDto(dto);
-
-        Assert.False(state.IsAllPupilsSelected);
-        Assert.Empty(state.GetSelectedPupils());
+        Assert.Throws<ArgumentException>(() => state.UpsertPupilSelectionState(mixedUpns, It.IsAny<bool>()));
     }
 }
